Upgrade deps, simplify app (#2)
There have been two libraries that have changed which have added
extra capability which we are upgrading to here. This also allows us
to remove the prometheus resource (which is now in the library,
along with health and so forth).

The move to Grizzly is because of a Netty/Jersey bug that is
causing large payloads (such as the metrics) to fail.